remove first tax with holding

+++ app/services/transaction_service.rb
@@ -61,7 +61,7 @@ class TransactionService
         remaining_weekend_price = (saturday * weekly_hours[:saturday_hours] + sunday * weekly_hours[:sunday_hours]) * listing.weekend_price.to_f
         remaining_weekly_earning = remaining_weekday_price + remaining_weekend_price
         weekly_tax_withholding = if tx.is_withholding_tax
-          @params[:transaction][:tax_withholding_amount].to_i.abs
+          @params[:transaction][:tax_withholding_amount].to_f.abs
         else
           0
         end
